Revisão de Funções

+++ src/main/java/com/enosnery/RestAPI/controllers/PriceTableController.java
@@ -6,15 +6,15 @@ import com.enosnery.RestAPI.interfaces.PriceTableDateGroup;
 import com.enosnery.RestAPI.interfaces.PriceTableDistributorGroup;
 import com.enosnery.RestAPI.models.PriceTableItem;
 import com.enosnery.RestAPI.services.PriceTableService;
-import com.enosnery.RestAPI.utils.CSVImport;
 import com.google.gson.Gson;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.fileupload.FileUploadException;
+import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.text.SimpleDateFormat;
 import java.util.HashMap;
 import java.util.List;
@@ -30,16 +30,30 @@ public class PriceTableController {
     @ApiOperation(value = "Lista de todos os itens da tabela de preços.", produces = "application/json", httpMethod = "GET")
     @RequestMapping(value = "/prices/list/all", method = RequestMethod.GET, produces = "application/json")
     public String prices(){
-        String json = new Gson().toJson(priceTableService.findAllItems());
-        return json;
+        HashMap<String, PriceTableItem> hashMap = new HashMap<>();
+        List<PriceTableItem> list = priceTableService.findAllItems();
+        if(list.size() == 0){
+            return "Não existe nenhum dado registrado";
+        }
+        for(PriceTableItem pti : list) {
+            hashMap.put("Item", pti);
+        }
+        return new Gson().toJson(hashMap);
     }
 
     @ApiOperation(value = "Salvar novo item na tabela de preços.", httpMethod = "POST")
     @PostMapping(value = "/prices/insert")
-    public Long insertItem(@RequestBody PriceTableItem item){
+    public String insertItem(@RequestBody PriceTableItem item){
         PriceTableItem newItem = new PriceTableItem(item.getRegion(), item.getState(), item.getCity(), item.getDistributor(), item.getInstallationCode(), item.getProduct(), item.getCollectDate(), item.getPurchasePrice(), item.getSalePrice(), item.getMeasurement(), item.getFlag());
         priceTableService.savePriceTableItem(newItem);
-        return newItem.getId();
+        if(newItem.getId() == null || newItem.getId() == 0){
+            return "Não foi possível salvar o item.";
+        }
+        HashMap<String, Long> map = new HashMap<>();
+        String sb = "itemId";
+        map.put(sb, newItem.getId());
+        return new Gson().toJson(map);
+
     }
 
     @ApiOperation(value = "Atualizar item na tabela de preços.", httpMethod = "POST")
@@ -48,7 +62,7 @@ public class PriceTableController {
         PriceTableItem updateItem = priceTableService.findById(id);
         if(updateItem != null) {
             priceTableService.savePriceTableItem(updateItem);
-            return updateItem.getId() != null ? "Atualizado!" : "Problemas ao atualizar.";
+            return updateItem.getId() != null ? "Item Atualizado!" : "Problemas ao atualizar o item.";
         }else{
             return "Esse item não existe.";
         }
@@ -57,29 +71,57 @@ public class PriceTableController {
     @ApiOperation(value = "Remover item da tabela de preços.", httpMethod = "POST")
     @PostMapping(value = "/prices/delete")
     public String deleteItem (@RequestParam Long id){
+        if(id == null || id == 0){
+            return "ID Inválido. Favor verificar e tentar novamente.";
+        }
         priceTableService.deleteItem(id);
         return "Item deletado.";
     }
 
     @ApiOperation(value = "Média de preço de venda de combustível dado uma cidade específica.", httpMethod = "GET")
     @RequestMapping(value = "/prices/average/city", method = RequestMethod.GET, produces = "application/json")
-    public Double averageByCity(@RequestParam String city){
-        return priceTableService.averageByCity(city);
+    public String averageByCity(@RequestParam String city){
+        if(city == null || city.equals("") || city.isEmpty()){
+            return "Preencha os campos corretamente.";
+        }
+        Double result = priceTableService.averageByCity(city);
+        if(result == null || result == 0d){
+            return "Cidade inválida";
+        }
+        HashMap<String, Double> map = new HashMap<>();
+        StringBuilder sb = new StringBuilder();
+        sb.append("Média da Cidade: ");
+        sb.append(city);
+        map.put(sb.toString(), result);
+        return new Gson().toJson(map);
     }
 
 
     @RequestMapping(value = "/prices/list/region", method = RequestMethod.GET, produces = "application/json")
     public String findByRegion(@RequestParam String region){
-        return new Gson().toJson(priceTableService.findByRegion(region));
+        HashMap<String, PriceTableItem> hashMap = new HashMap<>();
+        if(region == null || region.equals("") || region.isEmpty()){
+            return "Preencha os campos corretamente.";
+        }
+        List<PriceTableItem> list = priceTableService.findByRegion(region);
+        if(list.size() == 0){
+            return "Não existem registros para esta região";
+        }
+        for(PriceTableItem pti : list) {
+            hashMap.put("Item", pti);
+        }
+        return new Gson().toJson(hashMap);
     }
 
     @RequestMapping(value = "/prices/list/group", method = RequestMethod.GET, produces = "application/json")
     public String groupByDistributor(){
         HashMap<String, Integer> map = new HashMap<>();
         List<PriceTableDistributorGroup> list = priceTableService.groupByDistributor();
+        if(list.size() == 0){
+            return "Não foi possível agrupar os dados";
+        }
         for(PriceTableDistributorGroup ll : list){
             map.put(ll.getDistributor(), ll.getCount());
-
         }
         return new Gson().toJson(map);
     }
@@ -88,6 +130,9 @@ public class PriceTableController {
     public String groupByDate(){
         HashMap<String, Integer> map = new HashMap<>();
         List<PriceTableDateGroup> list = priceTableService.groupByDate();
+        if(list.size() == 0){
+            return "Não foi possível agrupar os dados.";
+        }
         for(PriceTableDateGroup dd : list){
             SimpleDateFormat format = new SimpleDateFormat("dd-MM-yyyy");
             String date = format.format(dd.getDate());
@@ -98,22 +143,56 @@ public class PriceTableController {
     }
 
     @RequestMapping(value = "/prices/list/city", method = RequestMethod.GET, produces = "application/json")
-    public List<PriceTableAverageCity> groupAverageByCity(){
+    public String groupAverageByCity(){
+        HashMap<String, HashMap> map = new HashMap<>();
+
         List<PriceTableAverageCity> list = priceTableService.groupAveragesByCity();
-        return list;
+        if(list == null || list.size() == 0){
+            return "Nenhum valor a ser mostrado";
+        }
+
+        for(PriceTableAverageCity ptac : list){
+            HashMap<String, Double> mapItem = new HashMap<>();
+            mapItem.put("SaleAverage", ptac.getSaleAverage());
+            mapItem.put("PurchaseAverage", ptac.getPurchaseAverage());
+            map.put(ptac.getCity(), mapItem);
+        }
+        return new Gson().toJson(map);
     }
 
     @RequestMapping(value = "/prices/list/flag", method = RequestMethod.GET, produces = "application/json")
-    public List<PriceTableAverageFlag> groupAverageByFlag(){
+    public String groupAverageByFlag(){
+        HashMap<String, HashMap> map = new HashMap<>();
         List<PriceTableAverageFlag> list = priceTableService.groupAveragesByFlag();
-        return list;
+        if(list == null || list.size() == 0){
+            return "Nenhum valor a ser mostrado";
+        }
+        for(PriceTableAverageFlag ptaf : list){
+            HashMap<String, Double> mapItem = new HashMap<>();
+            mapItem.put("SaleAverage", ptaf.getSaleAverage());
+            mapItem.put("PurchaseAverage", ptaf.getPurchaseAverage());
+            map.put(ptaf.getFlag(), mapItem);
+        }
+        return new Gson().toJson(map);
     }
 
     @PostMapping(value = "/prices/import", consumes = {"multipart/form-data"})
     public String importCSV(@RequestPart("file") MultipartFile file) throws Exception {
-        List<String[]> list = priceTableService.readAll(file);
-        priceTableService.saveAll(list);
-        return "CSV importado!";
+        try {
+            String extension = FilenameUtils.getExtension(file.getOriginalFilename());
+            if (extension != null && !(extension.equals("csv"))) {
+                return "Arquivo em Formato Inválido";
+            }
+            List<String[]> list = priceTableService.readAll(file);
+            if (list.size() == 0) {
+                return "Não foi possível importar os dados do CSV";
+            }
+            priceTableService.saveAll(list);
+            return "CSV importado!";
+        }catch (FileUploadException fe){
+            fe.printStackTrace();
+            return "Não foi adicionado nenhum arquivo.";
+        }
     }
 
 }
