+++ src/main/java/com/enosnery/RestAPI/controllers/UserController.java
@@ -3,8 +3,11 @@ package com.enosnery.RestAPI.controllers;
 import com.enosnery.RestAPI.forms.LoginForm;
 import com.enosnery.RestAPI.models.User;
 import com.enosnery.RestAPI.services.UserService;
+import com.google.gson.Gson;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.HashMap;
 
 @RestController
 public class UserController {
@@ -13,16 +16,29 @@ public class UserController {
     UserService userService;
 
     @PostMapping(value = "/login", produces = "application/json")
-    public Long login(@RequestBody LoginForm request){
-        return userService.findByLoginAndPassword(request.login, request.password);
+    public String login(@RequestBody LoginForm request){
+        if(request.login == null || request.login.equals("") || request.password == null || request.password.equals("")){
+            return "Preencha os campos corretamente!";
+        }
+        User user = userService.findByLoginAndPassword(request.login, request.password);
+        if(user == null || user.getId() == null || user.getId() == 0){
+            return "Usuário e/ou Senha incorretos";
+        }
+        return new Gson().toJson(user);
 
     }
 
     @PostMapping(value = "/user/insert")
-    public Long insertUser(@RequestBody User request){
+    public String insertUser(@RequestBody User request){
         User user = new User(request.getName(), request.getLogin(), request.getPassword());
         userService.saveUser(user);
-        return user.getId();
+        if(user.getId() == null || user.getId() == 0){
+            return "Não foi possível salvar o item.";
+        }
+        HashMap<String, Long> map = new HashMap<>();
+        String sb = "UserId";
+        map.put(sb, user.getId());
+        return new Gson().toJson(map);
     }
 
     @PostMapping(value = "/user/update")
@@ -38,6 +54,9 @@ public class UserController {
 
     @PostMapping(value = "/user/delete")
     public String deleteUser(@RequestParam Long id){
+        if(id == 0){
+            return "Preencha os campos corretamente";
+        }
         userService.deleteUser(id);
         return "Usuário deletado.";
     }
